+++ applications/backend/src/main/java/de/wohnungshelden/applications/Application.java
@@ -1,4 +1,4 @@
-package searchrequests;
+package de.wohnungshelden.applications;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
