+++ applications/backend/src/main/java/de/wohnungshelden/applications/HelloWorldController.java
@@ -1,4 +1,4 @@
-package searchrequests;
+package de.wohnungshelden.applications;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.RequestMapping;
